+++ web/web_common/web_common_traits/src/database/read/dispatch/async.rs
@@ -44,8 +44,8 @@ pub trait AsyncBoundedReadDispatch<C: AsyncConnection>: Sized {
     /// * Returns an error if loading the row fails.
     fn bounded_read(
         table_name: Self::TableName,
-        offset: u64,
-        limit: u64,
+        offset: u16,
+        limit: u16,
         conn: &mut C,
     ) -> impl Future<Output = Result<Self, diesel::result::Error>>;
 }
