+++ web/web_common/web_common_traits/src/database/read/sync/postgres.rs
@@ -34,8 +34,8 @@ where
     <<T::Table as OffsetDsl>::Output as LimitDsl>::Output: for<'a> LoadQuery<'a, PgConnection, T>,
 {
     fn bounded_read(
-        offset: u64,
-        limit: u64,
+        offset: u16,
+        limit: u16,
         conn: &mut PgConnection,
     ) -> Result<Vec<Self>, diesel::result::Error> {
         RunQueryDsl::load(
